test(x11): isolate guardian sequence contracts
